+++ ddl/task_user.sql
@@ -1,6 +1,6 @@
 CREATE TABLE TASK_USER(
-  task_id int not null references TASK(id) on delete cascade,
-  user_id int not null references APP_USER(id) on delete cascade,
+  task_id bigint not null references TASK(id) on delete cascade,
+  user_id bigint not null references APP_USER(id) on delete cascade,
   created_at timestamp not null default current_timestamp,
   updated_at timestamp not null default current_timestamp,
   primary key (task_id, user_id)
